refactor(model): disable SSR for model component, better separation of concerns
Server-side rendering caused the page to remain blank while the model was being initialised, which
was an issue with slower internet connections. The model now loads dynamically and only after the
rest of the page has loaded. Refactored model classes for better separation of concerns. Fixed low
model quality on mobile devices by adjusting the device pixel ratio of the renderer to match that of
the client device.